MAGETWO-50121: Fix configurable links

+++ app/code/Magento/ConfigurableProduct/Model/Product/SaveHandler.php
@@ -57,7 +57,10 @@ class SaveHandler
             return $entity;
         }
 
-        $this->deleteConfigurableProductAttributes($entity);
+        if ($extensionAttributes->getConfigurableProductOptions() !== null) {
+            $this->deleteConfigurableProductAttributes($entity);
+        }
+
         $configurableOptions = (array) $extensionAttributes->getConfigurableProductOptions();
         if (!empty($configurableOptions)) {
             $this->saveConfigurableProductAttributes($entity, $configurableOptions);
